jwt 인증 받도록 수정 및 연결

+++ backend/src/main/java/com/chakak/controller/ReportController.java
@@ -1,6 +1,7 @@
 package com.chakak.controller;
 
 import java.security.Principal;
+import java.util.Collections;
 import java.util.List;
 
 import org.springframework.data.domain.Page;
@@ -24,6 +25,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.chakak.domain.Report;
 import com.chakak.dto.ReportDto;
 import com.chakak.dto.request.ReportRequest;
+import com.chakak.service.CustomUserDetails;
 import com.chakak.service.ReportImageService;
 import com.chakak.service.ReportService;
 import lombok.RequiredArgsConstructor;
@@ -37,9 +39,14 @@ public class ReportController {
 	private final ReportImageService reportImageService;
 	
 	// ✅ 제보 신청 내역 저장 
-	/*@PostMapping
-	public ResponseEntity<?> saveReport(@RequestBody ReportRequest reportDto, Principal principal){
-		String userId = principal.getName(); 
+	@PostMapping
+	public ResponseEntity<?> saveReport(@RequestBody ReportRequest reportDto,  @AuthenticationPrincipal CustomUserDetails userDetails){
+		if (userDetails == null) {
+			 return ResponseEntity.status(401)
+			            .body(Collections.singletonMap("message", "로그인이 필요합니다."));
+	        }
+
+	     String userId = userDetails.getUsername();
 		
 		Report report = new Report();
 		report.setTitle(reportDto.getTitle());
@@ -53,27 +60,8 @@ public class ReportController {
 		
 		Report savedReport = reportService.save(report);
 		return ResponseEntity.ok(savedReport.getReportId());
-	}*/
-	
-	/// 임시 테스트용 다시 되돌릴 거임 
-	@PostMapping
-	public ResponseEntity<?> saveReport(@RequestBody ReportRequest reportDto) {
-	    String userId = "test1234";  //🌟🌟 하드코딩된 userId
-	    
-	    Report report = new Report();
-	    report.setTitle(reportDto.getTitle());
-	    report.setUserId(userId);
-	    report.setViolationType(reportDto.getViolationType());
-	    report.setVehicleNumber(reportDto.getVehicleNumber());
-	    report.setDescription(reportDto.getDescription());
-	    report.setAddress(reportDto.getAddress());
-	    report.setLatitude(reportDto.getLatitude());
-	    report.setLongitude(reportDto.getLongitude());
-	    
-	    Report savedReport = reportService.save(report);
-	    return ResponseEntity.ok(savedReport.getReportId());
 	}
-//////////////
+	
 	
 	// ✅ 제보 신청 내역(첨부 이미지) 저장 
 	@PostMapping("/upload/{reportId}")
@@ -89,45 +77,39 @@ public class ReportController {
 	
 	
 	// ✅ 전체 신고 목록 조회 or 필터링 조회 ( 차량 번호 , 위치 , 상태 , 글 쓴 날짜 , 기간 ( startDate , endDate) , 키워드 )
-	@GetMapping
-	public ResponseEntity<Page<ReportDto>> getAllReports(
-			// required = false는 뒤에 쿼리 파라미터가 안 붙이면 전체 목록 조회 , 붙이면 해당 파라미터 조회를 의미하는 것임
-	    @RequestParam(required = false) String carNumber, // RequestParam 이므로 ?로 붙이는 쿼리 파라미터임 
-	    @RequestParam(required = false) String location,
-	    @RequestParam(required = false) String reportDate,
-	    @RequestParam(required = false) String violationType,
-	    @RequestParam(required = false) String startDate,
-	    @RequestParam(required = false) String endDate,
-	    @RequestParam(required = false) String keyword,
-	    @PageableDefault(size = 10) Pageable pageable) {
+		@GetMapping
+		public ResponseEntity<Page<ReportDto>> getAllReports(
+				// required = false는 뒤에 쿼리 파라미터가 안 붙이면 전체 목록 조회 , 붙이면 해당 파라미터 조회를 의미하는 것임
+		    @RequestParam(required = false) String carNumber, // RequestParam 이므로 ?로 붙이는 쿼리 파라미터임 
+		    @RequestParam(required = false) String location,
+		    @RequestParam(required = false) String reportDate,
+		    @RequestParam(required = false) String violationType,
+		    @RequestParam(required = false) String startDate,
+		    @RequestParam(required = false) String endDate,
+		    @RequestParam(required = false) String keyword,
+		    @PageableDefault(size = 10) Pageable pageable) {
+			
 		
-		String userId = "test1234"; // 🌟🌟 하드코딩함
 
-	    Page<ReportDto> page = reportService.getAllReports(
-	        carNumber, location, 
-	        reportDate, violationType, startDate, endDate,
-	        keyword, pageable
-	    );
+		    Page<ReportDto> page = reportService.getAllReports(
+		        carNumber, location, 
+		        reportDate, violationType, startDate, endDate,
+		        keyword, pageable
+		    );
 
-	    return ResponseEntity.ok(page);
-	}
-	
+		    return ResponseEntity.ok(page);
+		}
+		
+		
 	
 	// ✅ 내 신고글 목록 조회하기 
-	/* @GetMapping("/my")
+	@GetMapping("/my")
 	    public ResponseEntity<List<ReportDto>> getMyReports(@AuthenticationPrincipal UserDetails userDetails) {
 	        String userId = userDetails.getUsername();
 	        List<ReportDto> reports = reportService.getMyReports(userId);
 	        return ResponseEntity.ok(reports);
-	    }*/
-	///테스트용 ////////////////////////////////
-	@GetMapping("/my")
-	public ResponseEntity<List<ReportDto>> getMyReports() {
-	    String userId = "test1234";  // 하드코딩 userId
-	    List<ReportDto> reports = reportService.getMyReports(userId);
-	    return ResponseEntity.ok(reports);
-	}
-	///////////////////////////////////////////
+	    }
+	
 	
 	// ✅ 상세 조회 ( 조회수 증가 ) 
 	 @GetMapping("/detail/{id}")
@@ -140,52 +122,49 @@ public class ReportController {
 	
 	  // ✅ 제보 신청 내역 수정
 	
-	@PutMapping("/{reportId}")
-	public ResponseEntity<?> updateReport(@PathVariable Long reportId, @RequestBody ReportRequest reportDto, Principal principal) {
-	    // 1. 수정할 대상 조회
-	    Report report = reportService.findById(reportId);
-	    if (report == null) {
-	        return ResponseEntity.notFound().build(); // 없는 경우 404 반환
-	    }
+	 @PutMapping("/{reportId}")
+		public ResponseEntity<?> updateReport(@PathVariable Long reportId, @RequestBody ReportRequest reportDto,  @AuthenticationPrincipal CustomUserDetails userDetails) {
+		    // 1. 수정할 대상 조회
+			if (userDetails == null) {
+	            return ResponseEntity.status(401).body("로그인이 필요합니다.");
+	        }
 
-	    // 2. 데이터 수정
-	    report.setTitle(reportDto.getTitle());
-	    //report.setUserId(principal.getName());
-	    report.setUserId("test1234"); // 🌟🌟 하드코딩 
-	    report.setViolationType(reportDto.getViolationType());
-	    report.setVehicleNumber(reportDto.getVehicleNumber());
-	    report.setDescription(reportDto.getDescription());
-	    report.setAddress(reportDto.getAddress());
-	    report.setLatitude(reportDto.getLatitude());
-	    report.setLongitude(reportDto.getLongitude());
+	        Report report = reportService.findById(reportId);
+	        if (report == null) {
+	            return ResponseEntity.notFound().build();
+	        }
 
-	    // 3. 저장
-	    Report updatedReport = reportService.save(report);
+		    // 2. 데이터 수정
+		    report.setTitle(reportDto.getTitle());
+		    report.setUserId(userDetails.getUsername());	 
+		    report.setViolationType(reportDto.getViolationType());
+		    report.setVehicleNumber(reportDto.getVehicleNumber());
+		    report.setDescription(reportDto.getDescription());
+		    report.setAddress(reportDto.getAddress());
+		    report.setLatitude(reportDto.getLatitude());
+		    report.setLongitude(reportDto.getLongitude());
 
-	    // 4. 결과 반환
-	    return ResponseEntity.ok(updatedReport.getReportId());
-	}
-	
+		    // 3. 저장
+		    Report updatedReport = reportService.save(report);
+
+		    // 4. 결과 반환
+		    return ResponseEntity.ok(updatedReport.getReportId());
+		}
+		
 	/**
 	 * 제보 신청 내역 삭제
 	 * */
-	@DeleteMapping("/{reportId}")
-	/*
-	public ResponseEntity<?> deleteReport(@PathVariable Long reportId, Principal principal) {
-		String userId = "test1234"; //🌟🌟 하드코딩함
-		reportService.deleteReport(reportId, principal.getName());
-	    return ResponseEntity.ok("제보가 삭제되었습니다.");
-	}*/
-	//🌟🌟🌟🌟🌟 테스트용 
-	
-	public ResponseEntity<?> deleteReport(@PathVariable Long reportId) {
-		String userId = "test1234"; // 테스트용
-		reportService.deleteReport(reportId, userId);
-		return ResponseEntity.ok("제보가 삭제되었습니다.");
-	}
-	///////////////////////
-	
-	
+	 @DeleteMapping("/{reportId}")
+		public ResponseEntity<?> deleteReport(@PathVariable Long reportId,  @AuthenticationPrincipal CustomUserDetails userDetails) {
+			 if (userDetails == null) {
+		            return ResponseEntity.status(401).body("로그인이 필요합니다.");
+		        }
+
+		        String userId = userDetails.getUsername();
+		        reportService.deleteReport(reportId, userId);
+		        return ResponseEntity.ok("제보가 삭제되었습니다.");
+		    }
+		
 	
 	
 }
